refactor: move Svg into SharedComponents
