+++ src/components/Accordion/Accordion.tsx
@@ -6,7 +6,7 @@ import Card from "@/components/SharedComponents/Card";
 import Flex from "@/components/SharedComponents/Flex";
 import Icon from "@/components/Icon";
 import Button from "@/components/SharedComponents/Button";
-import Svg from "@/components/Svg";
+import Svg from "@/components/SharedComponents/Svg";
 import BoxBorders from "@/components/SpriteSheet/BrushSvgs/BoxBorders";
 import Box from "@/components/SharedComponents/Box";
