+++ src/components/DownloadAndShareComponents/DownloadConfirmation/DownloadConfirmation.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 
 import Flex from "@/components/SharedComponents/Flex";
 import { Heading, P } from "@/components/SharedComponents/Typography";
-import Svg from "@/components/Svg";
+import Svg from "@/components/SharedComponents/Svg";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import NextLessonContainer from "@/components/DownloadComponents/NextLessonContainer";
 import { NextLesson } from "@/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.schema";
