+++ src/components/NewButtons/NewFocusUndeline.tsx
@@ -2,7 +2,8 @@ import styled, { css } from "styled-components";
 
 import { OakColorName } from "../../styles/theme";
 import getColorByName from "../../styles/themeHelpers/getColorByName";
-import Svg from "../Svg";
+
+import Svg from "@/components/SharedComponents/Svg";
 
 export const NewIconFocusUnderline = styled(Svg).attrs({
   name: "underline-1",
