+++ src/components/OakLink/FocusUnderline.tsx
@@ -2,7 +2,8 @@ import styled, { css } from "styled-components";
 
 import { OakColorName } from "../../styles/theme";
 import getColorByName from "../../styles/themeHelpers/getColorByName";
-import Svg from "../Svg";
+
+import Svg from "@/components/SharedComponents/Svg";
 
 const FocusUnderline = styled(Svg).attrs({ name: "underline-1" })<{
   $color: OakColorName;
