+++ src/components/OakLink/OakLink.tsx
@@ -12,13 +12,14 @@ import flex from "../../styles/utils/flex";
 import { OmitKeepDiscriminated } from "../../utils/generics";
 import { box } from "../SharedComponents/Box";
 import { HTMLAnchorProps } from "../SharedComponents/Button/common";
-import Svg from "../Svg";
+
 
 import FocusUnderline, { focusUnderlineStyles } from "./FocusUnderline";
 import NewFocusUnderline, {
   newFocusUnderlineStyles,
 } from "./NewFocusUnderline";
 
+import Svg from "@/components/SharedComponents/Svg";
 import { FlexProps } from "@/components/SharedComponents/Flex";
 
 type FocusStyle = "underline" | "new-underline";
