+++ src/components/SharedComponents/Card/CardLink.tsx
@@ -5,9 +5,10 @@ import Link from "next/link";
 import { OakLinkProps, transformOakLinkProps } from "../../OakLink";
 import { zIndexMap } from "../../../styles/utils/zIndex";
 import { DROP_SHADOW } from "../../../styles/utils/dropShadow";
-import Svg from "../../Svg";
 import getColorByName from "../../../styles/themeHelpers/getColorByName";
 import { HOVER_SHADOW_TRANSITION } from "../../../styles/transitions";
+
+import Svg from "@/components/SharedComponents/Svg";
 
 export const CardLinkFocusUnderline = styled(Svg).attrs((props) => ({
   ...props,
