+++ src/components/SharedComponents/Svg/Svg.stories.tsx
