+++ src/components/SharedComponents/Svg/index.ts
