+++ src/components/SharedComponents/Typography/Hr.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import Box, { BoxProps } from "../Box";
-import Svg from "../../Svg";
+import Box, { BoxProps } from "@/components/SharedComponents/Box";
+import Svg from "@/components/SharedComponents/Svg";
 
 type Thickness = 1 | 2 | 3 | 4 | 8;
