+++ src/components/SharedComponents/Typography/UnderlinedHeading.tsx
@@ -1,9 +1,10 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import Svg from "../../Svg";
-
 import Heading, { HeadingProps } from "./Heading";
+
+import Svg from "@/components/SharedComponents/Svg";
+
 
 const BrushUnderline = styled(Svg)`
   position: absolute;
