+++ src/components/SiteFooter/SiteFooter.tsx
@@ -5,12 +5,12 @@ import Logo from "../Logo";
 import SocialButtons from "../SocialButtons";
 import Grid, { GridArea } from "../Grid";
 import OakLink from "../OakLink";
-import Svg from "../Svg";
 import { OAK_SOCIALS } from "../SocialButtons/SocialButtons";
 import FooterSignpost from "../FooterSignpost/FooterSignpost";
 import { IconName } from "../Icon";
 import Icon from "../Icon/Icon";
 
+import Svg from "@/components/SharedComponents/Svg";
 import Typography, {
   Heading,
   LI,
