+++ src/components/SpriteSheet/BrushSvgs/BrushBorders/BrushBorders.tsx
@@ -2,10 +2,10 @@ import React, { FC } from "react";
 import styled, { css } from "styled-components";
 
 import getColorByName from "../../../../styles/themeHelpers/getColorByName";
-import Svg from "../../../Svg";
 import { OakColorName } from "../../../../styles/theme/types";
 import { getBreakpoint } from "../../../../styles/utils/responsive";
 
+import Svg from "@/components/SharedComponents/Svg";
 import { $selectedStyle } from "@/components/OakLink/OakLink";
 
 const brushBorder = css<{ color: OakColorName }>`
