+++ src/components/SpriteSheet/BrushSvgs/ButtonBorders/ButtonBorders.tsx
@@ -3,8 +3,9 @@ import styled, { css } from "styled-components";
 
 import getColorByName from "../../../../styles/themeHelpers/getColorByName";
 import { ButtonBackground } from "../../../SharedComponents/Button/common";
-import Svg from "../../../Svg";
 import { HOVER_SHADOW_TRANSITION } from "../../../../styles/transitions";
+
+import Svg from "@/components/SharedComponents/Svg";
 
 const buttonBorder = css<{ background: ButtonBackground }>`
   position: absolute;
