+++ src/components/SpriteSheet/BrushSvgs/ButtonBorders/DoubleButtonBorders.tsx
@@ -3,7 +3,7 @@ import styled, { css } from "styled-components";
 
 import getColorByName from "@/styles/themeHelpers/getColorByName";
 import { ButtonBackground } from "@/components/SharedComponents/Button/common";
-import Svg from "@/components/Svg";
+import Svg from "@/components/SharedComponents/Svg";
 
 const buttonBorder = css<{ background: ButtonBackground }>`
   position: absolute;
