+++ src/components/TagPromotional/TagPromotional.tsx
@@ -1,7 +1,6 @@
 import { FC } from "react";
 
-import Svg from "../Svg";
-
+import Svg from "@/components/SharedComponents/Svg";
 import { Span } from "@/components/SharedComponents/Typography";
 import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 import { OakColorName } from "@/styles/theme/types";
